apps/backend/src/app/auth/strategy/jwt.strategy.ts
import { UsersService } from '@autronas/backend/services';
import { Injectable } from '@nestjs/common';
import { PassportStrategy } from '@nestjs/passport';
import { ExtractJwt, Strategy } from 'passport-jwt';

@Injectable()
export class JwtStrategy extends PassportStrategy(Strategy) {
  constructor(protected readonly usersService: UsersService) {
    super({
      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
      ignoreExpiration: false,
      secretOrKey: process.env['JWT_SECRET'],
    });
  }

  async validate(payload: { id: string }) {
    const { id } = payload;

    const user = await this.usersService.getLogged(id);

    return user;
  }
}
